[bug] Remove the restore_repos() function
The restore_repos() function had the ability to completely destroy
a system if called at the wrong point. Instead of attempting to restore
the system, a fatal error is better.

I've also added a note in the README reminding folks that taking a full
backup before trying to switch the system is a very good idea.

Fixes #35.

Signed-off-by: Avi Miller <avi.miller@oracle.com>